move driver folder to component folder
